% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoards2.R
\name{get_boards2}
\alias{get_boards2}
\title{Title}
\usage{
get_boards2(
  board_width,
  board_height,
  pot_radius,
  pot_diameter,
  pot_rectangle_width,
  pot_rectangle_height,
  pots,
  pot_type,
  rectangle_enforce_given_orientation = FALSE
)
}
\arguments{
\item{board_width}{The width of a single board.
A board is defined as one of the 2 single elements used}

\item{board_height}{The height of a single board
A board is defined as one of the 2 single elements used}

\item{pot_radius}{The radius of a pot (for square and circular pots only)}

\item{pot_diameter}{The diameter of a pot (for square and circular pots only)}

\item{pot_rectangle_width}{The width of the pot, if it is a rectangular one.}

\item{pot_rectangle_height}{The height of the pot, if it is a rectangular one.}

\item{pots}{The number of pots you must place.}

\item{pot_type}{one of `c("circle","square","rectangle")`.}

\item{rectangle_enforce_given_orientation}{By default, the pot-fitting algorithm for
rectangular pots will try and fit as many pots onto a single board by testing
checking the result for the given pot_width/pot_height, and after flipping
the values for them, determining if flipping the pots by 90° will allow more
pots be fitted onto a single board.}
}
\value{
a 'board'-object containing
}
\description{
Title
}
\keyword{internal}
