% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptR.R
\docType{package}
\name{ptR}
\alias{ptR}
\alias{ptR-package}
\title{ptR - Process Trees in R}
\description{
This package provides function to analyse the behavior in a process tree, e.g. estimate the number of execution paths.
}

