% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTR_generateBoardLayouts2.R
\name{PTR_generateBoardLayouts2}
\alias{PTR_generateBoardLayouts2}
\title{Generate Board-Layouts for a number of pots.}
\usage{
PTR_generateBoardLayouts2(
  pots = NA,
  board_width = NA,
  board_height = NA,
  pot_radius = NA,
  pot_diameter = NA,
  pot_rectangle_width = NA,
  pot_rectangle_height = NA,
  distance = NA,
  pot_type = "circle",
  lbls = FALSE,
  rectangle_enforce_given_orientation = FALSE
)
}
\arguments{
\item{pots}{The number of pots you must place.}

\item{board_width}{The width of a single board.
A board is defined as one of the 2 single elements used}

\item{board_height}{The height of a single board
A board is defined as one of the 2 single elements used}

\item{pot_radius}{The radius of a pot (for square and circular pots only)}

\item{pot_diameter}{The diameter of a pot (for square and circular pots only)}

\item{pot_rectangle_width}{The width of the pot, if it is a rectangular one.}

\item{pot_rectangle_height}{The height of the pot, if it is a rectangular one.}

\item{distance}{The spacing distance between two pots}

\item{pot_type}{one of `c("circle","square","rectangle")`.}

\item{lbls}{A vector of length 'pos' consisting of custom names for each pot
Will be assigned in from bottom-left to top-right, from the first to the last
board.}

\item{rectangle_enforce_given_orientation}{By default, the pot-fitting algorithm for
rectangular pots will try and fit as many pots onto a single board by testing
checking the result for the given pot_width/pot_height, and after flipping
the values for them, determining if flipping the pots by 90° will allow more
pots be fitted onto a single board.}
}
\value{
layouts - list of layouts-objects, each containing a board-plot.
}
\description{
Given the dimensions of the pot, the number of required pots and the
dimensions of the shelf-boards, calculate the required number of occupied
boards. Optionally assign custom labels to them.
}
\examples{
# Example usage for rectangular pots
layouts_rect <- PTR_generateBoardLayouts2(
pots = 17,
board_width = 30,
board_height = 60,
pot_rectangle_width = 12,
pot_rectangle_height = 10,
distance = 0,
lbls = paste0("rectangle_", 1:17),
pot_type = "rectangle"
)
layouts_rect
# Example usage for circular pots
layouts_circle <- PTR_generateBoardLayouts2(
pots = 64,
board_width = 30,
board_height = 60,
pot_radius = 5,
distance = 0,
lbls = paste0("circle_", 1:64),
pot_type = "circle"
)
layouts_circle
# Example usage for square pots
layouts_square <- PTR_generateBoardLayouts2(
pots = 64,
board_width = 30,
board_height = 60,
pot_radius = 5,
pot_diameter = 5 * 2,
pot_rectangle_width = 5 * 2,
pot_rectangle_height = 5 * 2,
distance = 0,
lbls = paste0("square_", 1:64),
pot_type = "square"
)
layouts_square
}
