% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{PTR_load_dummy_data}
\alias{PTR_load_dummy_data}
\title{Generate a dummy BoardLayout as generated by `PTR_generateBoardLayouts` by feeding a set of dummy-data to it
This function is mostly supposed to be used for testing purposes, and should not be normally called.}
\usage{
PTR_load_dummy_data(set = 0)
}
\arguments{
\item{set}{Select a dummy testset and return the results
of `PTR_generateBoardLayouts`/`PTR_generateBoardLayouts2`.
As a rule, higher values for 'set' become more calculation-intense
Sets 1-3 leverage `PTR_generateBoardLayouts`
Sets 4-5 leverage `PTR_generateBoardLayouts2`}
}
\value{
Layouts as defined by `PTR_generateBoardLayouts`
}
\description{
Generate a dummy BoardLayout as generated by `PTR_generateBoardLayouts` by feeding a set of dummy-data to it
This function is mostly supposed to be used for testing purposes, and should not be normally called.
}
\examples{
layout2 <- PTR_load_dummy_data(2)
}
