% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting.R
\name{PTR_rotateBoards}
\alias{PTR_rotateBoards}
\title{Title}
\usage{
PTR_rotateBoards(boards, shifts = -1)
}
\arguments{
\item{boards}{boards to shift around}

\item{shifts}{steps to shift. Positive integers shift left-to-right, while positive integers shift right-to-left}
}
\value{
a new set of boards containing shifted pots
}
\description{
To shift the boards themselves, consider using `PTR_rotatePots`
}
\examples{
\dontrun{
oldBoards <- PTR_generateBoardLayouts(40, 20, 20, 5, 0)
newBoards <- PTR_rotateBoards(oldBoards, -2)
}
}
