% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting.R
\name{PTR_rotateBoards2}
\alias{PTR_rotateBoards2}
\title{rotate Boards forward or backwards through the list.}
\usage{
PTR_rotateBoards2(boards, shifts = -1)
}
\arguments{
\item{boards}{boards to shift around}

\item{shifts}{steps to shift. Positive integers shift a board to the next-up
position (last board becomes the first, first board becomes 2nd), while
negative integers move them backwards (e.g. the 2nd board becomes the first
board and the first board moves to the last position.}
}
\value{
a new set of boards containing shifted pots
}
\description{
To shift the pots themselves, consider using `PTR_rotatePots2`
}
\examples{
\dontrun{
ob2 <- PTR_load_dummy_data(5) # load example dummy data for the `2`-functions
nb21 <- PTR_rotateBoards2(ob2, -1) # board 1 goes to last position, board 2
# goes to first position
nb22 <- PTR_rotateBoards2(ob2, +1) # board 1 goes to 2nd position, last board
# goes to first position
}
}
\keyword{internal}
