% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting.R
\name{PTR_rotatePots}
\alias{PTR_rotatePots}
\title{Title}
\usage{
PTR_rotatePots(boards, shifts = 2)
}
\arguments{
\item{boards}{boards in which pots are to be shifted within}

\item{shifts}{steps to shift. Positive integers shift  clockwise,
while negative integers shift left-to-right,bottom-to-top}
}
\value{
boards with shifted pots. The boards themselves remain at their location.
}
\description{
To shift the boards themselves, consider using `PTR_rotateBoards`
}
\examples{
\dontrun{
oldBoards <- generateBoardLayouts(40, 20, 20, 5, 0)
newBoards <- rotatePots(oldBoards, -2)
}
}
