% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifting.R
\name{PTR_sortPots_by_potindex}
\alias{PTR_sortPots_by_potindex}
\title{Sort Labels of a board-object based on replicate-ID/replicate-**number**}
\usage{
PTR_sortPots_by_potindex(boards, shifts = -1)
}
\arguments{
\item{boards}{boards in which pots are to be shifted within}

\item{shifts}{steps to shift. Positive integers shift  clockwise,
while negative integers shift left-to-right,bottom-to-top}
}
\value{
boards with shifted pots. The boards themselves remain at their location.
}
\description{
Use this function after generating a board to create an ordered layout where
members of groups are equally distributed across all available boards.
}
\details{
NOTE:
This requires that the number is separable from the group-name
 by a single `_`.
}
\examples{
\dontrun{
boards_with_consecutive_groups <- PTR_load_dummy_data(4) # load example dummy data
# with custom group-names `c("UU","UG","ABAU","ABAG")`, each repeated 8
# times (`UU1,UU2,UU3,...UG1,UG2,...ABAU1,ABAU2,...ABAG1,...ABAG8`)
newBoards <- rotatePots(oldBoards, -2)
}
}
