% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTR_test_call.R
\name{PTR_test_call}
\alias{PTR_test_call}
\title{Generate a dummy BoardLayout as generated by `PTR_generateBoardLayouts` by feeding a set of dummy-data to it
This function is mostly supposed to be used for testing purposes, and should not be normally called.}
\usage{
PTR_test_call(set = 1)
}
\arguments{
\item{set}{Select a dummy testset and return the results of `PTR_generateBoardLayouts`. As a rule, higher values for 'set' become more calculation-intense}
}
\value{
Layouts as defined by `PTR_generateBoardLayouts`
}
\description{
Generate a dummy BoardLayout as generated by `PTR_generateBoardLayouts` by feeding a set of dummy-data to it
This function is mostly supposed to be used for testing purposes, and should not be normally called.
}
\examples{
layout2 <- PTR_test_call(2)
}
