% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{validate1}
\alias{validate1}
\title{validate inputs for function set 1}
\usage{
validate1(inputs)
}
\arguments{
\item{pots}{/}

\item{board_width}{/}

\item{board_height}{/}

\item{pot_radius}{/}

\item{pot_diameter}{/}

\item{distance}{/}

\item{lbls}{/}
}
\value{
/
}
\description{
validate inputs for function set 1
}
\keyword{internal}
