% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{validate2}
\alias{validate2}
\title{validate inputs for function set 2}
\usage{
validate2(inputs)
}
\arguments{
\item{pots}{/}

\item{board_width}{/}

\item{board_height}{/}

\item{pot_radius}{/}

\item{pot_diameter}{/}

\item{pot_rectangle_width}{/}

\item{pot_rectangle_height}{/}

\item{distance}{/}

\item{lbls}{/}

\item{pot_type}{/}
}
\value{
/
}
\description{
validate inputs for function set 2
}
\keyword{internal}
