% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTR_generateBoardLayouts.R
\name{PTR_generateBoardLayouts}
\alias{PTR_generateBoardLayouts}
\title{Generate Board-Layouts for a number of pots.}
\usage{
PTR_generateBoardLayouts(
  pots = NA,
  board_width = NA,
  board_height = NA,
  pot_radius = NA,
  distance = NA,
  lbls = FALSE
)
}
\arguments{
\item{pots}{The number of pots you must place}

\item{board_width}{The width of a single board.
A board is defined as one of the 2 single elements used}

\item{board_height}{The height of a single board.
See 'board_width' for details}

\item{pot_radius}{The radius of a pot}

\item{distance}{The spacing distance between two pots.}

\item{lbls}{vector of length 'pots' consisting of custom names for each pot.
Will be assigned in from bottom-left to top-right}
}
\value{
layouts - list of layout_objects, each containing a board-plot.
Additionally, the input data and a mapping of coordinates to pot-names
plotted is returned
}
\description{
Given the dimensions of the pot, the number of required pots and the
dimensions of the shelf-boards, calculate the required number of occupied
boards. Optionally assign custom labels to them.
}
\details{
WARNING: This function does not support rectangular pots. For that, use
`PTR_generateBoardLayouts2`
}
\examples{
#EXAMPLES
pots <- 34
board_width <- 31 # in [unit] - width of a board-section
board_height <- 30 # in [unit] - height of a board-section
pot_radius <- 3.625 # in [unit] - radius of a pot/the bounding circle/platter
distance <- 0 # in [unit] - spacing between adjacent pots
boardLayout <- PTR_generateBoardLayouts(pots, board_width, board_height, pot_radius, distance)
}
